+++ R/write.R
@@ -1,6 +1,37 @@
+
+#' Handle Missing Values in eeg_lst objects
+#'
+#' Remove all NAs from an `eeg_lst` signal table.
+#'
+#' @param object an `eeg_lst`.
+#' @param ... 	further arguments special methods could require.
+#' @export
+na.omit.eeg_lst <- function(object, ...){
+  object$.signal <- na.omit(object$.signal)
+  if (!is.null(object$.events) && nrow(object$.events) > 0) {
+    range_s <- object$.signal[, .(.lower = min(.sample), .upper = max(.sample)), by = .id]
+    object$.events <- update_events(object$.events, range_s)
+  }
+  object$.segments <- semi_join.(object$.segments, object[[1]], by = ".id")
+  object
+}
+
+#' Write an eeg_lst object to BrainVision file(s) (experimental)
+#'
+#' The function will write each recording in a different file.
+#'
+#' @param x `eeg_lst` object.
+#'
+#' @param file Either a vector of file names (with the same length as the number of recordings), one file name (which will be appended the recording name if needed), or a path which will use the recording names as file names. By default, it will use the recording name in the current path as file names.
+#' @param overwrite Stop writing if the file exists.
+#'
 #' @export
 write_vhdr <- function(x, file, overwrite = FALSE){
-  # do by recording
+  nr <- nrow(x$.signal)
+  # remove NAs
+  x <- na.omit(x)
+  nr_diff <- nr - nrow(x$.signal)
+  if(nr_diff>0) warning(nr_diff, " samples were removed because they contained NAs.", call. = FALSE)
  recs <- unique(x$.segments$.recording)
  if(length(file) ==1 & dir.exists(file)) {
    file <- file.path(file, recs)
@@ -9,14 +40,15 @@ write_vhdr <- function(x, file, overwrite = FALSE){
  } else if(length(file) ==1 & !dir.exists(file)) {
    file <- paste0(file, "_",recs)
  }
- if(overwrite == FALSE & any(file.exists(file))){
-   stop("write_vhdr cannot overwrite file",call. = FALSE)
- }
+
  for(n in seq_along(recs)){
-  xr <- x %>% eeg_filter(.recording==recs[n])
-  write_dat(xr, file[n])
-  write_vhdr_metadata(xr, file[n])
-  write_vmrk(xr, file[n])
+  xr <- x %>% eeg_filter(.recording == recs[n]) %>%
+    eeg_unsegment()
+  message_verbose("Saving file '",tools::file_path_sans_ext(file[n]),"{.dat,.vhdr,.vmrk}'...")
+  write_dat(xr, file[n], overwrite = overwrite)
+  write_vhdr_metadata(xr, file[n], overwrite = overwrite)
+   # make it work when it is segmented
+   write_vmrk(xr, file[n], overwrite = overwrite)
  }
 }
 
@@ -32,8 +64,9 @@ write_vhdr <- function(x, file, overwrite = FALSE){
 # }
 
 #' @noRd
-write_dat <- function(x, file){
+write_dat <- function(x, file, overwrite= FALSE){
   file <- paste0(tools::file_path_sans_ext(file),".dat")
+  if(file.exists(file) & overwrite==FALSE) stop("The file '",file, "' already exists.", call. = FALSE)
   dat <- file(file, "wb")
   on.exit(close(dat))
   #by default IEEE_FLOAT_32
@@ -44,8 +77,10 @@ write_dat <- function(x, file){
 
 
 #' @noRd
-write_vhdr_metadata <- function(x, file){
+write_vhdr_metadata <- function(x, file, overwrite= FALSE){
   fname <- tools::file_path_sans_ext(basename(file))
+  file <- paste0(tools::file_path_sans_ext(file),".vhdr")
+  if(file.exists(file) & overwrite==FALSE) stop("The file '",file, "' already exists.", call. = FALSE)
   ver2 <- paste0("Brain Vision Data Exchange Header File Version 2.0\n; Data created with the R package `eeguana` version ", packageVersion("eeguana"),"\n")
   chs <- channel_names(x)
   chs_n<- paste0("Ch",1:length(chs))
@@ -77,23 +112,28 @@ write_vhdr_metadata <- function(x, file){
   "<Resolution in \"Unit\">,<Unit>, Future extensions...",
   "Fields are delimited by commas, some fields might be omitted (empty).",
   "Commas in channel names are coded as \"\\1\".")
-  write.ini(vhdr, paste0(tools::file_path_sans_ext(file),".vhdr"),"UTF-8")
+  write.ini(vhdr, file,"UTF-8")
   
 }
 
 #' @noRd
-write_vmrk <- function(x, file) {
+write_vmrk <- function(x, file, overwrite= FALSE) {
   fname <- tools::file_path_sans_ext(basename(file))
+  file <- paste0(tools::file_path_sans_ext(file),".vmrk")
+  if(file.exists(file) & overwrite==FALSE) stop("The file '",file, "' already exists.", call. = FALSE)
   ver2 <- paste0("Brain Vision Data Exchange Marker File, Version 2.0\n; Data created with the R package `eeguana` version ", packageVersion("eeguana"),"\n")
   
   chs_n <- as.numeric(factor(x$.events$.channel,levels = channel_names(x)))
   chs_n <- ifelse(is.na(chs_n),0, chs_n)
-  # deal with several .ids in events
-  #  x$.events %>% split( f= ".id")
+  
   markers <- with(x$.events, paste0(.type,",", .description,",", .initial, ",", .final - .initial + 1,",",chs_n))
-  #this should be added only to new segment
-  # TO CHECK what happens in epoched datasets
-  markers[1] <- paste0(markers[1],",",chr_remove(format(Sys.time(), "%Y%m%d%H%M%OS6"),"\\."))
+  
+  for(n in seq_along(markers)){
+    if (chr_detect(markers[n],"New Segment")) {
+      markers[n] <- paste0(markers[n],",",chr_remove(format(Sys.time(), "%Y%m%d%H%M%OS6"),"\\."))
+    }
+  }
+  
   
   names(markers) <- paste0("Mk",1:length(markers))
   vmrk <- list("Common Infos" = list(Codepage= "UTF-8", 
@@ -102,10 +142,8 @@ write_vmrk <- function(x, file) {
                "Marker User Infos"= list()
   )
   attributes(vmrk)$title <- ver2
-  
-  write.ini(vmrk, paste0(tools::file_path_sans_ext(file),".vmrk"),"UTF-8")
-  
-  }
+  write.ini(vmrk, file,"UTF-8")
+}