+++ tests/testthat/test_01-reading.R
@@ -183,9 +183,14 @@ test_that("special vhdr file",{
 
 
 test_that("write vhdr",{
-  write_vhdr(multiplexed_bin_bv2,"test")
+  write_vhdr(x = multiplexed_bin_bv2,file = "test", overwrite = TRUE)
   multiplexed_bin_bv2_t <- read_vhdr("test.vhdr",.recording="bv2")
   expect_equal(multiplexed_bin_bv2_t,multiplexed_bin_bv2)
-  mul <- bind(multiplexed_bin_bv2,multiplexed_bin_bv2 %>% eeg_mutate(.recording = "bv3"))
-  
-})
+  mul <- bind(multiplexed_bin_bv2,multiplexed_bin_bv2 %>% eeg_mutate(Fp1 =Fp1*0.1, .recording = "bv3"))
+  write_vhdr(mul, "test_mul", overwrite = TRUE)
+  mul_test <- bind(read_vhdr("test_mul_bv2.vhdr",.recording = "bv2"),read_vhdr("test_mul_bv3.vhdr", .recording ="bv3"))
+  expect_equal(mul, mul_test,tolerance = 0.0001)
+  write_vhdr(x = seg_ascii_bv2,file="test_seg", overwrite = TRUE)
+  test_seg <- read_vhdr(file = "test_seg_bv2.vhdr", .recording = "bv2")
+  expect_equal(seg_ascii_bv2,test_seg,tolerance = 0.0001 )
+    })