+++ tests/testthat/test_06-artifacts.R
@@ -5,7 +5,7 @@ set.seed(123)
 N <- 1000
 signal <- dplyr::tibble(
   .id = 1L,
-  .sample = sample_int(seq_len(N), sampling_rate = 500),
+  .sample = sample_int(seq_len(N), .sampling_rate = 500),
   Fz = channel_dbl(rep(0, N)),
   Cz = channel_dbl(rep(0, N)),
   Pz = channel_dbl(rep(0, N))
@@ -23,7 +23,7 @@ data_1minmax <- eeg_lst(
   signal_tbl = dplyr::tibble(
     X = channel_dbl(c(0, 0, 0, -10, 10, 0, 0, 0, 0)),
     .id = 1L,
-    .sample = sample_int(1:9, sampling_rate = 500)
+    .sample = sample_int(1:9, .sampling_rate = 500)
   ),
   segments = dplyr::tibble(.id = 1L, .recording = "recording1", segment = 1L)
 )
@@ -32,7 +32,7 @@ data_1step <- eeg_lst(
   signal_tbl = dplyr::tibble(
     X = channel_dbl(c(0, 0, 0, 0, 10, 10, 10, 10, 10)),
     .id = 1L,
-    .sample = sample_int(1:9, sampling_rate = 500)
+    .sample = sample_int(1:9,.sampling_rate = 500)
   ),
   segments = dplyr::tibble(.id = 1L, .recording = "recording1", segment = 1L)
 )
@@ -42,7 +42,7 @@ data_more <- eeg_lst(
     signal %>%
       dplyr::mutate(
         .id = rep(1:4, each = N / 4),
-        .sample = sample_int(rep(seq_len(N / 4), times = 4), sampling_rate = 500)
+        .sample = sample_int(rep(seq_len(N / 4), times = 4), .sampling_rate = 500)
       ),
   segments = dplyr::tibble(.id = seq.int(4), .recording = paste0("recording", c(1, 1, 2, 2)), segment = seq.int(4))
 )