+++ tests/testthat/test_09-plots.R
@@ -7,7 +7,7 @@ data_1 <- eeg_lst(
     X = sin(1:30),
     Y = sin(1:30),
     .id = rep(c(1L, 2L, 3L), each = 10),
-    .sample = sample_int(rep(seq(-4L, 5L), times = 3), sampling_rate = 500)
+    .sample = sample_int(rep(seq(-4L, 5L), times = 3), .sampling_rate = 500)
   ),
   channels_tbl = dplyr::tibble(
     .channel = c("X", "Y"), .reference = NA, theta = NA, phi = NA,